% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{adresse_voie_prx}
\alias{adresse_voie_prx}
\title{adresse_voie_prx}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 60 rows and 1 columns.
}
\usage{
adresse_voie_prx
}
\description{
\preformatted{## Rows: 60
## Columns: 1
## $ lib_voie <chr> "(grande )?voie", "(rue du )?faubourg", "(rue du )?fg", "all", "all[eé]es?", "...
}
}
\keyword{datasets}
