% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{conv_code_postal}
\alias{conv_code_postal}
\title{Obtenir le code postal correspondant a une commune (a partir d'un code postal plus fin)}
\usage{
conv_code_postal(code_postal)
}
\arguments{
\item{code_postal}{Un vecteur de codes postaux de type caractère.}
}
\value{
Un vecteur de codes postaux correspondant à un code commune.

Jeu de données source : \code{geographie::data_ptt}.\cr
Il est créé à partir de la table "N_PTT" de la BCE.
}
\description{
Obtenir le code postal correspondant à une commune (à partir d'un code postal plus fin).
}
\examples{
geographie::conv_code_postal(c("75015", "75115"))

}

