% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{conv_commune_cp}
\alias{conv_commune_cp}
\title{Obtenir le code postal a partir d'un code commune}
\usage{
conv_commune_cp(code_commune)
}
\arguments{
\item{code_commune}{Un vecteur de code commune de type caractère.}
}
\value{
Un vecteur de code postaux.\cr
Seuls les codes postaux non-doublonnés sont retournés.

Jeu de données source : \code{geographie::ptt}.\cr
}
\description{
Obtenir le code postal à partir d'un code commune.
}
\examples{
geographie::conv_commune_cp(c("18110", "38061"))

}
