% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{conv_cp_commune}
\alias{conv_cp_commune}
\title{Obtenir le code commune a partir d'un code postal}
\usage{
conv_cp_commune(code_postal)
}
\arguments{
\item{code_postal}{Un vecteur de codes postaux de type caractère.}
}
\value{
Un vecteur de code commune.\cr
Seuls les codes communes non-doublonnés sont retournés.

Jeu de données source : \code{geographie::ptt}.\cr
Il est créé à partir de la table "N_PTT" de la BCE.
}
\description{
Obtenir le code commune à partir d'un code postal.
}
\examples{
# Deux exemples de codes postaux : le premier n'est lié qu'à un seul code commune mais pas le second
geographie::conv_cp_commune(c("01001", "01000"))

}
