% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{conv_cp_ville_commune}
\alias{conv_cp_ville_commune}
\title{Obtenir le code commune a partir d'un code postal et d'un libelle de commune}
\usage{
conv_cp_ville_commune(code_postal, lib_commune)
}
\arguments{
\item{code_postal}{Un vecteur de codes postaux de type caractère.}

\item{lib_commune}{Un vecteur de libellés de commune.}
}
\value{
Un vecteur de code commune.

Jeu de données source : \code{geographie::data_cp_ville_commune}.\cr
Il est créé à partir du jeu de données ODS Référentiel géographique français (voir projet "Géographie").
}
\description{
Obtenir le code commune à partir d'un code postal et d'un libellé de commune.
}
\examples{
geographie::conv_cp_ville_commune(code_postal = c("17000", "34010"),
  lib_commune = c("La Rochelle", "Montpellier"))

}
