% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{conv_lib_code_pays}
\alias{conv_lib_code_pays}
\title{Obtenir le code pays INSEE a partir de libelle de pays}
\usage{
conv_lib_code_pays(lib_pays)
}
\arguments{
\item{lib_pays}{Un vecteur de libellés de pays.

Utilise less tables "Pays" et "Pays_libelle" de la base Access "Tables_ref.accdb".}
}
\value{
Un vecteur de code pays INSEE.
}
\description{
Obtenir le code pays INSEE à partir de libellés de pays.
}
\examples{
geographie::conv_lib_code_pays(c("france", "Etats-Unis"))

}
