% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cp_ville_commune}
\alias{cp_ville_commune}
\title{cp_ville_commune}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 80979 rows and 3 columns.
}
\usage{
cp_ville_commune
}
\description{
\preformatted{## Rows: 80,979
## Columns: 3
## $ code_postal  <chr> "209", "211", "213", "217", "218", "221", "222", "223", "224", "225", "226...
## $ lib_commune  <chr> "PARIS 12", "PARIS 12", "PARIS 12", "PARIS 12", "PARIS 12", "PARIS 12", "P...
## $ code_commune <chr> "75112", "75112", "75112", "75112", "75112", "75112", "75112", "75112", "7...
}
}
\keyword{datasets}
