% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detecter_adresse.R
\name{decouper_adresse_lignes}
\alias{decouper_adresse_lignes}
\title{Decouper une adresse complete en ligne}
\usage{
decouper_adresse_lignes(adresse)
}
\arguments{
\item{adresse}{Un vecteur d'adresses.}
}
\value{
Une liste de vecteurs (taille de liste équivalente à la taille de \code{adresse}).\cr
Chaque élément de liste contient un vecteur dont la taille correspond au nombre de segment de l'adresse.

Le découpage est réalisé selon trois critères :\cr
 - Le caractère de saut de ligne\cr
 - La présence d'un libellé de type voie renseigné dans la table \code{geographie::adresse_voie_prx}\cr
 - La présence d'une boite postale (BP) ou d'une course spéciale (CS)
}
\description{
Découper une adresse complète en ligne.
}
\examples{
geographie::decouper_adresse_lignes(c("MENESR 1 Rue descartes", "1 Rue descartes BP 1"))

}
