% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{departement}
\alias{departement}
\title{departement}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 111 rows and 4 columns.
}
\usage{
departement
}
\description{
\preformatted{## Rows: 111
## Columns: 4
## $ code_departement <chr> "000", "001", "002", "003", "004", "005", "006", "007", "008", "009", ...
## $ lib_departement  <chr> "Armées", "Ain", "Aisne", "Allier", "Alpes-de-Haute-Provence", "Hautes...
## $ code_region      <chr> NA, "84", "32", "84", "93", "93", "93", "84", "44", "76", "44", "76", ...
## $ code_region_2015 <chr> NA, "82", "22", "83", "93", "93", "93", "82", "21", "73", "21", "91", ...
}
}
\keyword{datasets}
