% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocoder_adresse.R
\name{geocoder_adresse}
\alias{geocoder_adresse}
\title{Obtenir le resultat des API (adresse.data.gouv.fr et google) a partir de l'adresse complete}
\usage{
geocoder_adresse(adresse, service, nettoyer_adresse = TRUE, timeout = 10,
  cle_google = NULL)
}
\arguments{
\item{adresse}{Un vecteur d'adresses.}

\item{service}{L'API à utiliser: \code{adresse.data.gouv.fr} ou \code{googlemaps}.}

\item{nettoyer_adresse}{\code{TRUE}, l'adresse est nettoyée (minuscule, sans accent et ponctuation); \code{FALSE}, aucune opération n'est réalisée.}

\item{timeout}{Temps de réponse maximal en secondes.}
}
\value{
Un data_frame avec une ligne par adresse soumise.\cr

Les résultats des géocodages sont accumulés et sauvegardés :\cr
- Dans \code{geographie::data_geocodage_data_gouv} pour le service \code{adresse.data.gouv.fr}\cr
- Dans \code{geographie::data_geocodage_google} pour le service \code{googlemaps}\cr

Pour le service \code{adresse.data.gouv.fr}, le data frame contient les champs suivants :\cr
- \code{adresse} : adresse initiale\cr
- \code{adresse_nettoyee} : adresse nettoyée\cr
- \code{date_ajout} : date de téléchargement\cr
(Les champs suivants sont les éléments retournés par l'API)\cr
- \code{match}\cr
- \code{score}\cr
- \code{type}\cr
- \code{latitude}\cr
- \code{longitude}\cr
- \code{code_postal}\cr
- \code{code_commune}\cr
- \code{type_localisation}\cr
- \code{numero_voie}\cr
- \code{lib_voie}\cr
- \code{localite}\cr

Pour le service \code{googlemaps}, le data frame contient les champs suivants :\cr
- \code{adresse} : adresse initiale\cr
- \code{adresse_nettoyee} : adresse nettoyée\cr
- \code{date_ajout} : date de téléchargement\cr
(Les champs suivants sont les éléments retournés par l'API)\cr
- \code{latitude}\cr
- \code{longitude}\cr
- \code{type_localisation}\cr
- \code{statut}\cr
}
\description{
Obtenir le résultat des API (adresse.data.gouv.fr et google) à partir de l'adresse complète.
}
\examples{
# Exemple avec le service adresse.gouv pour le MENESR et Rennes 2
geographie::geocoder_adresse(c("1 Rue Descartes 75005 Paris", "Place Recteur Henri le Moal 35000 Rennes"),
  service = "adresse.data.gouv.fr")

# Exemple avec le service google pour le MENESR et Rennes 2
geographie::geocoder_adresse(c("1 Rue Descartes 75005 Paris", "Place Recteur Henri le Moal 35000 Rennes"),
  service = "googlemaps")
}
