% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocoder_adresse.R
\name{geocoder_adresse_data_gouv}
\alias{geocoder_adresse_data_gouv}
\title{Geocoder des adresses a partir de l'API adresse.data.gouv.fr}
\usage{
geocoder_adresse_data_gouv(adresse, timeout = 10)
}
\arguments{
\item{adresse}{Un vecteur d'adresses.}

\item{timeout}{Temps de réponse maximal en secondes.}

\item{nettoyer_adresse}{\code{TRUE}, l'adresse est nettoyée (minuscule, sans accent et ponctuation); \code{FALSE}, aucune opération n'est réalisée.}
}
\value{
A matrix of the infile
}
\description{
Géocoder des adresses à partir de l'API adresse.data.gouv.fr.
}
\keyword{internal}

