% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocoder_adresse.R
\name{geocoder_adresse_google}
\alias{geocoder_adresse_google}
\title{Geocoder des adresses a partir de l'API google}
\usage{
geocoder_adresse_google(adresse, timeout = 10)
}
\arguments{
\item{adresse}{Un vecteur d'adresses.}

\item{timeout}{Temps de réponse maximal en secondes.}
}
\value{
A matrix of the infile
}
\description{
Géocoder des adresses à partir de l'API google.
}
\keyword{internal}
