% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchie.R
\name{hier_commune_departement}
\alias{hier_commune_departement}
\title{Obtenir le code departement a partir d'un code commune}
\usage{
hier_commune_departement(code_commune)
}
\arguments{
\item{code_commune}{Un vecteur de code commune.}
}
\value{
Un vecteur de code de département.

Jeu de données source : \code{geographie::data_ods_geo}.\cr
Il est créé à partir du jeu de données ODS Référentiel géographique français (voir projet "Géographie").
}
\description{
Obtenir le code département à partir d'un code commune.
}
\examples{
geographie::hier_commune_departement(c("01001", "33003"))

}
