% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchie.R
\name{hier_commune_region}
\alias{hier_commune_region}
\title{Obtenir le code region a partir d'un code commune}
\usage{
hier_commune_region(code_commune)
}
\arguments{
\item{code_commune}{Un vecteur de code commune.}
}
\value{
Un vecteur de code de région.

Jeu de données source : \code{geographie::ods_geo}.\cr
}
\description{
Obtenir le code région à partir d'un code commune.
}
\examples{
geographie::hier_commune_region(c("01001", "33003"))

}
