% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchie.R
\name{hier_commune_region_2015}
\alias{hier_commune_region_2015}
\title{Obtenir le code region (2015 et avant) a partir d'un code commune}
\usage{
hier_commune_region_2015(code_commune)
}
\arguments{
\item{code_commune}{Un vecteur de code commune.}
}
\value{
Un vecteur de code de région (2015 et avant).

Jeu de données source : \code{geographie::ods_geo}.\cr
Il est créé à partir du jeu de données ODS Référentiel géographique français (voir projet "Géographie").
}
\description{
Obtenir le code région (2015 et avant) à partir d'un code commune.
}
\examples{
geographie::hier_commune_region_2015(c("01001", "33003"))

}
