% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchie.R
\name{hier_departement_region_2015}
\alias{hier_departement_region_2015}
\title{Obtenir le code region (2015 et avant) a partir d'un code departement}
\usage{
hier_departement_region_2015(code_departement)
}
\arguments{
\item{code_commune}{Un vecteur de code département.}
}
\value{
Un vecteur de code de région (2015 et avant).

Jeu de données source : \code{geographie::departement}.\cr
}
\description{
Obtenir le code région (2015 et avant) à partir d'un code département.
}
\examples{
geographie::hier_departement_region_2015(c("003", "056"))

}
