% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libelle.R
\name{lib_commune}
\alias{lib_commune}
\title{Obtenir le libelle de commune a partir d'un code commune}
\usage{
lib_commune(code_commune)
}
\arguments{
\item{code_commune}{Un vecteur de code commune.}
}
\value{
Un vecteur de libellé de commune.

Jeu de données source : \code{geographie::ods_geo}.\cr
}
\description{
Obtenir le libellé de commune à partir d'un code commune.
}
\examples{
geographie::lib_commune(c("01001", "33003"))

# Création d'un champ dans un data frame avec la fonction "mutate"
data <- dplyr::tibble(code_commune = c("01001", "33003"))
dplyr::mutate(data, libelle = geographie::lib_commune(code_commune))

}
