% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libelle.R
\name{lib_pays}
\alias{lib_pays}
\title{Obtenir le libelle de pays a partir d'un code pays (code INSEE)}
\usage{
lib_pays(code_pays, langue = "fr")
}
\arguments{
\item{code_pays}{Un vecteur de code pays (code INSEE).}

\item{langue}{Un code langue (\code{fr} ou \code{en}).}
}
\value{
Un vecteur de libellé de pays.

Jeu de données source : \code{geographie::data_ods_geo}.\cr
Il est créé à partir de la table "Pays" de la base Access "Tables_ref.accdb" (voir projet "Géographie").
}
\description{
Obtenir le libellé de pays à partir d'un code pays (code INSEE).
}
\examples{
geographie::lib_pays(c("100", "109"))

# Création d'un champ dans un data frame avec la fonction "mutate"
dplyr::data_frame(code_pays = c("100", "109")) \%>\%
  dplyr::mutate(libelle = geographie::lib_pays(code_pays),
                libelle_en = geographie::lib_pays(code_pays, langue = "en"))

}

