% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libelle.R
\name{lib_type_voie}
\alias{lib_type_voie}
\title{Obtenir le libelle de type de voie a partir du code}
\usage{
lib_type_voie(code_type_voie)
}
\arguments{
\item{code_type_voie}{Un vecteur de code de type de voie.}
}
\value{
Un vecteur de libellé de type de voie.

Jeu de données source : \code{geographie::data_type_voie}.\cr
Il est créé à partir de la table "Adresse_voie_type" de la base Access "Tables_ref.accdb" (voir projet "Géographie").
}
\description{
Obtenir le libellé de type de voie à partir du code.
}
\examples{
geographie::lib_type_voie(c("AV", "BD", "QUA"))

# Création d'un champ dans un data frame avec la fonction "mutate"
dplyr::data_frame(code_type_voie = c("AV", "BD", "QUA")) \%>\%
  dplyr::mutate(libelle = geographie::lib_type_voie(code_type_voie))

}

