% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detecter_adresse.R
\name{localiser_adresse}
\alias{localiser_adresse}
\title{Localiser une adresse dans une chaine de caracteres}
\usage{
localiser_adresse(adresse, regex_adresse)
}
\arguments{
\item{adresse}{Un vecteur d'adresses.}

\item{regex_adresse}{Une expression régulière.}
}
\value{
Un vecteur numérique contenant la position de départ de l'adresse.
}
\description{
Localiser une adresse dans une chaine de caractères.
}
\keyword{internal}
