% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ods_geo}
\alias{ods_geo}
\title{ods_geo}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 38916 rows and 35 columns.
}
\usage{
ods_geo
}
\description{
\preformatted{## Rows: 38,916
## Columns: 35
## $ regrgp_nom        <chr> "Province", "Province", "Province", "Province", "Province", "Province...
## $ lib_region        <chr> "Auvergne-Rhône-Alpes", "Auvergne-Rhône-Alpes", "Auvergne-Rhône-Alpes...
## $ lib_region_2015   <chr> "Rhône-Alpes", "Rhône-Alpes", "Rhône-Alpes", "Provence-Alpes-Côte d'A...
## $ aca_nom           <chr> "Lyon", "Lyon", "Lyon", "Aix-Marseille", "Aix-Marseille", "Aix-Marsei...
## $ lib_departement   <chr> "Ain", "Ain", "Ain", "Alpes-de-Haute-Provence", "Alpes-de-Haute-Prove...
## $ code_commune      <chr> "01030", "01032", "01009", "04034", "04086", "04084", "04051", "04082...
## $ com_code1         <chr> "01030", "01032", "01009", "04034", "04086", "04084", "04051", "04082...
## $ com_code2         <chr> "01030", "01032", "01009", "04034", "04086", "04084", "04051", "04082...
## $ com_id            <chr> "C01030", "C01032", "C01009", "C04034", "C04086", "C04084", "C04051",...
## $ com_nom_maj_court <chr> "BEAUREGARD", "BELIGNEUX", "ANDERT ET CONDON", "LA BRILLANNE", "FAUCO...
## $ com_nom_maj       <chr> "BEAUREGARD", "BELIGNEUX", "ANDERT-ET-CONDON", "LA BRILLANNE", "FAUCO...
## $ lib_commune       <chr> "Beauregard", "Béligneux", "Andert-et-Condon", "La Brillanne", "Fauco...
## $ code_uu           <chr> "00758", "01201", NA, "04201", NA, NA, NA, NA, NA, NA, NA, NA, NA, "0...
## $ uu_id             <chr> "UU00758", "UU01201", "SO", "UU04201", "SO", "SO", "SO", "SO", "SO", ...
## $ uucr_id           <chr> "UU00758", "UU01201", "CR01009", "UU04201", "CR04086", "CR04084", "CR...
## $ lib_uu            <chr> "Lyon", "Béligneux", "Andert-et-Condon", "Oraison", "Faucon-de-Barcel...
## $ ze_id             <chr> "ZE8213", "ZE8214", "ZE8216", "ZE9302", "ZE9304", "ZE9301", "ZE9301",...
## $ code_departement  <chr> "01", "01", "01", "04", "04", "04", "04", "04", "04", "04", "04", "04...
## $ dep_id            <chr> "D001", "D001", "D001", "D004", "D004", "D004", "D004", "D004", "D004...
## $ dep_nom_num       <chr> "Ain (01)", "Ain (01)", "Ain (01)", "Alpes-de-Haute-Provence (04)", "...
## $ dep_num_nom       <chr> "01 - Ain", "01 - Ain", "01 - Ain", "04 - Alpes-de-Haute-Provence", "...
## $ aca_code          <chr> "10", "10", "10", "02", "02", "02", "02", "02", "02", "02", "02", "02...
## $ aca_id            <chr> "A10", "A10", "A10", "A02", "A02", "A02", "A02", "A02", "A02", "A02",...
## $ code_region       <chr> "84", "84", "84", "93", "93", "93", "93", "93", "93", "93", "93", "93...
## $ reg_id            <chr> "R84", "R84", "R84", "R93", "R93", "R93", "R93", "R93", "R93", "R93",...
## $ code_region_2015  <chr> "82", "82", "82", "93", "93", "93", "93", "93", "93", "93", "93", "93...
## $ reg_id_old        <chr> "R82", "R82", "R82", "R93", "R93", "R93", "R93", "R93", "R93", "R93",...
## $ fd_id             <chr> "FD111", "FD111", "FD111", "FD111", "FD111", "FD111", "FD111", "FD111...
## $ fr_id             <chr> "FR11", "FR11", "FR11", "FR11", "FR11", "FR11", "FR11", "FR11", "FR11...
## $ fe_id             <chr> "FE1", "FE1", "FE1", "FE1", "FE1", "FE1", "FE1", "FE1", "FE1", "FE1",...
## $ uu_id_99          <chr> "UU00553", "SO", "SO", "UU04103", "SO", "SO", "SO", "SO", "SO", "SO",...
## $ au_code           <chr> "002", "002", "294", "519", "670", "218", NA, NA, NA, NA, NA, NA, NA,...
## $ au_id             <chr> "AU002", "AU002", "AU294", "AU519", "AU670", "AU218", "AU998", "SO", ...
## $ auc_id            <chr> "AU002", "AU002", "AU294", "AU519", "AU670", "AU218", "C04051", "SO",...
## $ auc_nom           <chr> "Lyon", "Lyon", "Belley", "Oraison", "Barcelonnette", "Digne-les-Bain...
}
}
\keyword{datasets}
