% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pays}
\alias{pays}
\title{pays}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 254 rows and 8 columns.
}
\usage{
pays
}
\description{
\preformatted{## Rows: 254
## Columns: 8
## $ code_pays       <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA,...
## $ lib_pays_fr     <chr> "Îles Turques-et-Caïques", "Saint-Martin (Partie Française)", "Îles Mar...
## $ lib_pays_en     <chr> "Turks and Caicos Islands", "Saint Martin (French part)", "Northern Mar...
## $ code_langue     <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA,...
## $ code_pays_iso2  <chr> "TC", "MF", "MP", "MS", "NC", "NF", "NU", "PM", "PU", "KY", "SJ", "PF",...
## $ code_pays_iso3  <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA,...
## $ article_pays    <chr> "les", NA, "les", "le", "la", NA, NA, NA, NA, "les", NA, "la", "les", "...
## $ lib_nationalite <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA,...
}
}
\keyword{datasets}
