% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pays_libelle}
\alias{pays_libelle}
\title{pays_libelle}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 73 rows and 2 columns.
}
\usage{
pays_libelle
}
\description{
\preformatted{## Rows: 73
## Columns: 2
## $ libelle_pays <chr> "angleterre", "azerbajan", "bolivia plurinational state of", "bonaire sain...
## $ code_pays    <chr> "132", "253", "418", "443", "118", "132", "331", "323", "404", "312", "312...
}
}
\keyword{datasets}
