% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ptt}
\alias{ptt}
\title{ptt}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 59347 rows and 14 columns.
}
\usage{
ptt
}
\description{
\preformatted{## Rows: 59,347
## Columns: 14
## $ ligne_acheminement_net <chr> "ARMEES", "ARMEES", "ARMEES", "ARMEES", "ARMEES", "ARMEES", "ARM...
## $ nom_localite_net       <chr> "PARIS", "PARIS", "PARIS", "PARIS", "PARIS", "PARIS", "PARIS", "...
## $ particularite_commune  <chr> "C", "C", "C", "C", "C", "C", "C", "C", "C", "C", "C", "C", "C",...
## $ nom_localite           <chr> "PARIS 12", "PARIS 12", "PARIS 12", "PARIS 12", "PARIS 12", "PAR...
## $ nom_localite_38        <chr> "PARIS 12", "PARIS 12", "PARIS 12", "PARIS 12", "PARIS 12", "PAR...
## $ code_postal            <chr> "209", "211", "213", "217", "218", "221", "222", "223", "224", "...
## $ bureau_distributeur    <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, ...
## $ ligne_acheminement     <chr> "ARMEES", "ARMEES", "ARMEES", "ARMEES", "ARMEES", "ARMEES", "ARM...
## $ ligne_acheminement_32  <chr> "ARMEES", "ARMEES", "ARMEES", "ARMEES", "ARMEES", "ARMEES", "ARM...
## $ id                     <dbl> 57900, 57901, 57902, 58218, 58219, 58220, 58221, 58222, 58223, 5...
## $ created_at             <dbl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, ...
## $ updated_at             <dbl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, ...
## $ n_commentaire          <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, ...
## $ code_commune           <chr> "75112", "75112", "75112", "75112", "75112", "75112", "75112", "...
}
}
\keyword{datasets}
