% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{region}
\alias{region}
\title{region}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 41 rows and 5 columns.
}
\usage{
region
}
\description{
\preformatted{## Rows: 41
## Columns: 5
## $ code_region      <chr> "00", "01", "02", "03", "04", "06", "07", "08", "11", "21", "22", "23"...
## $ lib_region       <chr> "TOM + Collectivités territoriales", "Guadeloupe", "Martinique", "Guya...
## $ ajout_rd         <chr> NA, NA, NA, NA, NA, NA, "O", "O", NA, NA, NA, NA, NA, NA, NA, NA, NA, ...
## $ latitude_region  <dbl> NA, 16.261774, 14.638664, 3.941459, -21.133571, -12.831664, NA, NA, 48...
## $ longitude_region <dbl> NA, -61.560559, -60.985478, -53.186451, 55.551885, 45.158413, NA, NA, ...
}
}
\keyword{datasets}
