% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{type_voie}
\alias{type_voie}
\title{type_voie}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 39 rows and 2 columns.
}
\usage{
type_voie
}
\description{
\preformatted{## Rows: 39
## Columns: 2
## $ code_type_voie <chr> "ALL", "AV", "BD", "CAR", "CHE", "CHS", "CITE", "COR", "CRS", "DOM", "ES...
## $ lib_type_voie  <chr> "Allée", "Avenue", "Boulevard", "Carrefour", "Chemin", "Chaussée", "Cité...
}
}
\keyword{datasets}
