% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{conv_pays_eu_insee}
\alias{conv_pays_eu_insee}
\title{Obtenir le code pays INSEE a partir du code pays EU}
\usage{
conv_pays_eu_insee(code_pays_eu)
}
\arguments{
\item{code_pays_eu}{Un vecteur de code pays EU.

Jeu de données source : \code{geographie::data_pays}.\cr
Il est créé à partir de la table "Pays" de la base Access "Tables_ref.accdb" (voir projet "Géographie").}
}
\value{
Un vecteur de code pays INSEE.
}
\description{
Obtenir le code pays INSEE à partir du code pays EU.
}
\examples{
geographie::conv_pays_eu_insee(c("FR", "DE"))

}

