% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detecter_adresse.R
\name{extraire_bp_cs}
\alias{extraire_bp_cs}
\title{Extraire d'une adresse la boite postale et/ou la course speciale (CS)}
\usage{
extraire_bp_cs(adresse)
}
\description{
Extraire d'une adresse la boite postale et/ou la course spéciale (CS).
}
\keyword{internal}
